# gen/openapi/v1.yaml
components:
  schemas:
...
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    ListNodesOutputBody:
      additionalProperties: false
      properties:
        nodes:
          items:
            $ref: "#/components/schemas/Node"
          type: array
      required:
        - nodes
      type: object
...
    Node:
      additionalProperties: false
      properties:
        announcedServices:
          description: The services in the node's serve configuration, as it last reported them.
          items:
            $ref: "#/components/schemas/NodeService"
          type: array
        appConnector:
          description: true while the client runs the app connector service (tailscale set --advertise-connector).
          type: boolean
        approved:
          description: false while the node waits for an administrator.
          type: boolean
        approvedAt:
          format: date-time
          type:
            - string
            - "null"
        approvedRoutes:
          items:
            type: string
          type: array
        approvedServices:
          description: The services the node may host; it hosts the ones it also announces.
          items:
            type: string
          type: array
        availableRoutes:
          items:
            type: string
          type: array
        clientVersion:
          description: The Tailscale client version the node last reported, such as 1.86.2; empty until it connects.
          type: string
        clientWarnings:
          description: "Problems the client reports about itself: ip-forwarding-off for a subnet router whose kernel drops forwarded packets, router-unhealthy for a broken route setup, etc-apt-source-disabled when the Tailscale apt source is commented out. A newer client may report flags not listed here. Empty while the client reports none, while it is offline, and after a restart of the server until it polls again."
          items:
            type: string
          type: array
        createdAt:
          format: date-time
          type: string
        discoKey:
          type: string
        ephemeral:
          description: true when the node is deleted on logout or after the ephemeral timeout.
          type: boolean
        exitNodePriority:
          description: "Order among the global exit nodes for clients that pick one automatically: the highest wins, 0 is no preference."
          format: int64
          type: integer
        expiry:
          format: date-time
          type:
            - string
            - "null"
        funnelEnabled:
          description: true while the client has a Funnel endpoint on, exposing a service to the internet through the ingress.
          type: boolean
        givenName:
          type: string
        globalExitNode:
          description: true when every client is told to prefer this exit node.
          type: boolean
        hardwareAttestation:
          $ref: "#/components/schemas/NodeHardwareAttestation"
          description: What the machine's hardware attestation key proved; absent until the client signs a map request with one.
        id:
          format: uint64
          type: string
        ipAddresses:
          items:
            type: string
          type: array
        lastSeen:
          format: date-time
          type:
            - string
            - "null"
        machineKey:
          type: string
        name:
          type: string
        netInfo:
          $ref: "#/components/schemas/NodeNetInfo"
          description: The client's last network report; absent until it connects.
        nodeKey:
          type: string
        online:
          type: boolean
        os:
          description: The operating system the client reported, as Tailscale names it (linux, macOS, windows, iOS, android, freebsd); empty until it connects.
          type: string
        osVersion:
          description: The operating system version the client reported, such as 15.1 or Ubuntu 24.04; empty until it connects.
          type: string
        preAuthKey:
          $ref: "#/components/schemas/NodePreAuthKey"
        registerMethod:
          enum:
            - REGISTER_METHOD_UNSPECIFIED
            - REGISTER_METHOD_AUTH_KEY
            - REGISTER_METHOD_CLI
            - REGISTER_METHOD_OIDC
          type: string
        remoteConfig:
          description: true while the client delegated remote configuration to the control plane (tailscale set --remote-config).
          type: boolean
        sharedWith:
          description: IDs of the users the node is shared with.
          items:
            type: string
          type: array
        sshServer:
          description: true while the client runs Tailscale SSH (tailscale set --ssh).
          type: boolean
        subnetRoutes:
          items:
            type: string
          type: array
        suspended:
          description: true while an administrator has suspended the node.
          type: boolean
        suspendedAt:
          format: date-time
          type:
            - string
            - "null"
        tags:
          items:
            type: string
          type: array
        tpm:
          $ref: "#/components/schemas/NodeTPM"
          description: The TPM the client found; absent when it reported none.
        updateAvailable:
          description: true when a newer stable Tailscale client exists than the one the node runs; see latestClientVersion on the server info.
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - machineKey
        - nodeKey
        - discoKey
        - ipAddresses
        - name
        - user
        - lastSeen
        - expiry
        - preAuthKey
        - createdAt
        - approved
        - approvedAt
        - suspended
        - suspendedAt
        - registerMethod
        - givenName
        - online
        - approvedRoutes
        - availableRoutes
        - subnetRoutes
        - tags
        - sharedWith
        - globalExitNode
        - exitNodePriority
        - announcedServices
        - approvedServices
        - funnelEnabled
        - clientVersion
        - updateAvailable
        - os
        - osVersion
        - ephemeral
        - appConnector
        - sshServer
        - clientWarnings
        - remoteConfig
      type: object
...
    NodeDERPLatency:
      additionalProperties: false
      properties:
        code:
          type: string
        ipv4Ms:
          description: Over IPv4, 0 when not measured.
          format: double
          type: number
        ipv6Ms:
          description: Over IPv6, 0 when not measured.
          format: double
          type: number
        ms:
          description: The fastest recent round trip in milliseconds, over IPv4 or IPv6.
          format: double
          type: number
        name:
          type: string
        regionId:
          format: int64
          type: integer
      required:
        - regionId
        - code
        - name
        - ms
        - ipv4Ms
        - ipv6Ms
      type: object
    NodeHardwareAttestation:
      additionalProperties: false
      properties:
        attested:
          description: true when the last map request carried a valid signature by the key; it is the node:hardwareAttested posture attribute.
          type: boolean
        attestedAt:
          description: When attestation was last gained; it is not refreshed per request.
          format: date-time
          type:
            - string
            - "null"
        key:
          description: The key that last verified, as hwattestpub:<hex>.
          type: string
        keyChangedAt:
          description: When a signature last arrived under a new key; null while it never did.
          format: date-time
          type:
            - string
            - "null"
      required:
        - attested
        - attestedAt
        - keyChangedAt
        - key
      type: object
    NodeNetInfo:
      additionalProperties: false
      properties:
        havePortMap:
          description: Whether a UPnP, NAT-PMP or PCP mapping is open.
          type: boolean
        latency:
          items:
            $ref: "#/components/schemas/NodeDERPLatency"
          type: array
        linkType:
          description: wired, wifi or mobile, when the client knows.
          type: string
        mappingVariesByDestIp:
          description: "true behind a hard NAT: a mapping per destination."
          type:
            - boolean
            - "null"
        pcp:
          description: Whether PCP was seen on the LAN.
          type:
            - boolean
            - "null"
        pmp:
          description: Whether NAT-PMP was seen on the LAN.
          type:
            - boolean
            - "null"
        preferredDerp:
          description: The relay region the client homes on, 0 while unknown.
          format: int64
          type: integer
        preferredDerpName:
          description: That region's name from the relay map.
          type: string
        upnp:
          description: Whether UPnP was seen on the LAN.
          type:
            - boolean
            - "null"
        workingIpv6:
          description: Whether the client reaches the internet over IPv6.
          type:
            - boolean
            - "null"
        workingUdp:
          description: Whether UDP reaches the internet.
          type:
            - boolean
            - "null"
      required:
        - preferredDerp
        - preferredDerpName
        - latency
        - mappingVariesByDestIp
        - workingIpv6
        - workingUdp
        - havePortMap
        - upnp
        - pmp
        - pcp
        - linkType
      type: object
...
    NodePreAuthKey:
      additionalProperties: false
      properties:
        aclTags:
          items:
            type: string
          type: array
        createdAt:
          format: date-time
          type:
            - string
            - "null"
        ephemeral:
          type: boolean
        expiration:
          format: date-time
          type:
            - string
            - "null"
        id:
          format: uint64
          type: string
        key:
          type: string
        preauthorized:
          type: boolean
        reusable:
          type: boolean
        used:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
          description: Absent for a tagged key.
      required:
        - id
        - key
        - reusable
        - ephemeral
        - used
        - expiration
        - createdAt
        - aclTags
        - preauthorized
      type: object
...
    NodeService:
      additionalProperties: false
      properties:
        active:
          description: true when the node advertises the service.
          type: boolean
        name:
          description: The service name, svc:<label>.
          type: string
        ports:
          description: The protocol and ports the node serves it on.
          items:
            type: string
          type: array
      required:
        - name
        - ports
        - active
      type: object
...
    NodeTPM:
      additionalProperties: false
      properties:
        firmwareVersion:
          description: The firmware version.
          format: int64
          minimum: 0
          type: integer
        manufacturer:
          description: The four-letter manufacturer code, such as MSFT.
          type: string
        model:
          description: The vendor-defined model.
          format: int64
          type: integer
        specRevision:
          description: The TPM 2.0 specification revision.
          format: int64
          type: integer
        vendor:
          description: The vendor string.
          type: string
      required:
        - manufacturer
        - vendor
        - model
        - firmwareVersion
        - specRevision
      type: object
...
    User:
      additionalProperties: false
      properties:
        approved:
          description: false while the user waits for an administrator.
          type: boolean
        approvedAt:
          format: date-time
          type:
            - string
            - "null"
        createdAt:
          format: date-time
          type: string
        displayName:
          type: string
        email:
          type: string
        id:
          format: uint64
          type: string
        name:
          type: string
        profilePicUrl:
          type: string
        provider:
          type: string
        providerId:
          type: string
        role:
          description: owner, admin, network-admin, it-admin, auditor or member
          type: string
      required:
        - id
        - name
        - createdAt
        - displayName
        - email
        - providerId
        - provider
        - profilePicUrl
        - role
        - approved
        - approvedAt
      type: object
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http
info:
  description: Slopscale control server API.
  title: Slopscale API
  version: v1
openapi: 3.1.0
paths:
...
  /api/v1/node:
    get:
      description: A credential with devices:core:read lists every node. Any other credential owned by a user lists the nodes that user owns and the ones shared with them, so a member sees their own machines; a credential without a user sees none.
      operationId: listNodes
      parameters:
        - explode: false
          in: query
          name: user
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListNodesOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      security:
        - bearer: []
      summary: List nodes
      tags:
        - Nodes
...